Initial success at detecting edges in a frame.
Initial success at detecting edges in a frame.
The edge image is really busy though with tons of edges. Not sure if this function needs to be tuned or if the image needs to be tuned before being passed to it. TBD.

Also, getting some weird messages in the console while running, despite the fact that it is working. These will need to be investigated